requirements.txt
stomp.py~=6.0.0
Flask~=1.1.2
aiohttp~=3.6.2
uWSGI==2.0.18
